+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.7 ##
+
+1.  Fix `MinMaxSlider` incorrect step.
+2.  Fix `ProgressBar` set value is not in real-time in struct type.
+
 ## 2.1.6 ##
 
 1.  UI Toolkit: Fix `MinMaxSlider` incorrect clamp.